+++ build.gradle
@@ -21,6 +21,8 @@ dependencies {
 
     implementation 'net.rakugakibox.spring.boot:logback-access-spring-boot-starter:2.7.1'
     implementation 'pl.allegro.tech.boot:handlebars-spring-boot-starter:0.3.1'
+    implementation 'org.springframework.boot:spring-boot-starter-websocket'
+
 
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'junit:junit:4.12'
