+++ src/main/java/chess/domain/User.java
@@ -1,0 +1,4 @@
+package chess.domain;
+
+public class User {
+}
