+++ src/main/java/chess/domain/game/Position.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.game;
 
 import java.util.*;
