+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,6 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.Position;
+import chess.domain.game.Position;
 
 import java.util.List;
 import java.util.Map;
