+++ src/main/java/chess/domain/team/WhiteTeam.java
@@ -1,6 +1,6 @@
 package chess.domain.team;
 
-import chess.domain.Position;
+import chess.domain.game.Position;
 import chess.domain.piece.Piece;
 
 import java.util.Map;
