+++ src/main/java/chess/service/room/ChessRoomService.java
@@ -1,6 +1,5 @@
 package chess.service.room;
 
-import chess.domain.Room;
 import dto.ChessGameDto;
 import dto.RoomDto;
 import dto.RoomRequestDto;
