+++ src/main/java/dto/PiecesDto.java
@@ -1,6 +1,6 @@
 package dto;
 
-import chess.domain.Position;
+import chess.domain.game.Position;
 import chess.domain.piece.Piece;
 import util.PieceConverter;
