+++ src/test/java/chess/domain/piece/PawnTest.java
@@ -1,8 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.Position;
-import chess.domain.piece.Pawn;
-import chess.domain.piece.Piece;
+import chess.domain.game.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
