+++ src/test/java/chess/domain/team/BlackTeamTest.java
@@ -1,8 +1,7 @@
 package chess.domain.team;
 
-import chess.domain.Position;
+import chess.domain.game.Position;
 import chess.domain.piece.*;
-import chess.domain.team.BlackTeam;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
